Mantis.Oracle

Description

Oracle for general data.

Documentation

makeValidator

Arguments

:: Oracle
The oracle.

-> Data
The datum.

-> Action
The redeemer.

-> ScriptContext
The context.

-> Bool
Whether the transaction is valid.

Make the validator for the oracle.

data OracleScript

Type for the script.

oracleInstance

Arguments

:: Oracle
The oracle.

-> ScriptInstance OracleScript
The instance.

Compute the instance for an oracle.

oracleValidator

Arguments

:: Oracle
The oracle.

-> Validator
The validator.

Compute the validator for an oracle.

oracleAddress

Arguments

:: Oracle
The oracle.

-> Address
The script address.

Compute the address for an oracle.

findOracle

Arguments

:: HasBlockchainActions s

=> Oracle
The oracle.

-> Contract w s Text (Maybe (TxOutRef, TxOutTx, Data))
Action for finding the oracle's UTxO and datum.

Find an oracle on the blockchain.

fetchDatum

Arguments

:: TxOut
The transaction output.

-> (DatumHash -> Maybe Datum)
Function for looking up the datum, given its hash.

-> Maybe Data
The datum, if it was found.

Retrieve the oracle's datum from a transaction output.

exportOracle

Arguments

:: FilePath
The filename for writing the validator bytes.

-> Oracle
The oracle.

-> IO Address
Action writing the validator and returning its address.

Export the validator for an oracle and compute its address.

Orphan instances

IsData Action

Instance details

Methods

toData :: Action -> Data

fromData :: Data -> Maybe Action

Lift DefaultUni Oracle

Instance details

Methods

lift :: Oracle -> RTCompile DefaultUni fun (Term TyName Name DefaultUni fun ())

Typeable DefaultUni Oracle

Instance details

Methods

typeRep :: Proxy Oracle -> RTCompile DefaultUni fun (Type TyName DefaultUni ())