Mantis.Oracle.Simulate

Description

Example usage for the oracle.

Documentation

main

Arguments

:: CurrencySymbol
Currency symbol for the example tokens.

-> TokenName
Name of the control token.

-> TokenName
Name of the datum token.

-> TokenName
Name of the fee token.

-> Integer
Amount of fee token needed to read the oracle.

-> IO ()
Action for running the example.

Run the example.

testTrace

Arguments

:: Parameters
The parameters for the oracle.

-> EmulatorTrace ()
The action for running the test.

Trace the oracles operation for a given set of parameters.

peekDatum

Arguments

:: Oracle
The oracle.

-> Contract () BlockchainActions Text ()
Action for logging the oracle's data.

Log the oracle's current data.

peekFunds

Arguments

:: Oracle
The oracle.

-> Maybe Integer
The wallet number, or Nothing for the oracle script.

-> Contract (Last Value) BlockchainActions Text ()
Action for logging the funds.

Log the funds in a wallet or in the oracle script.