Mantis.Oracle.Simulate.PAB

Description

Plutus application backend simulation for oracle contracts.

Documentation

data TheContracts

Contract specifications.

Constructors

Init TokenName TokenName TokenName Integer [(Wallet, Integer, FilePath)]
Initializer with control token name, datum token name, fee token name, fee amount, and wallets with initial fee tokens and CID path.

TheController Parameters
Controller.

TheClient Oracle
Client.

runPAB

Arguments

:: TokenName
Name for the control token.

-> TokenName
Name for the datum token.

-> TokenName
Name for the fee token.

-> Integer
The fee amount.

-> [(Wallet, Integer, FilePath)]
Wallets, their initial fee token amount, and the paths to their Contract ID (CID) files.

-> IO ()
Action for running the PAB.

Run the Plutus application backend. The first wallet controls the oracle.

waitForLast

Arguments

:: FromJSON a

=> ContractInstanceId
The Contract ID.

-> Simulation t a
Action for waiting.

Wait for the success of a contract.

handleTheContracts :: Member (Error PABError) effs => Member (LogMsg (PABMultiAgentMsg (Builtin TheContracts))) effs => ContractEffect (Builtin TheContracts) ~> Eff effs

Handle the contracts.

handlers :: SimulatorEffectHandlers (Builtin TheContracts)

Handlers for the contract.

initContract

Arguments

:: TokenName
Name for the control token.

-> TokenName
Name for the datum token.

-> TokenName
Name for the fee token.

-> Integer
The fee amount.

-> [(Wallet, Integer, FilePath)]
Wallets, their initial fee token amount, and the paths to their Contract ID (CID) files.

-> Contract (Last Parameters) BlockchainActions Text ()
Action for initializing the contract.

Initialize the contract. The first wallet controls the oracle.

snd3 :: (a, b, c) -> b

Extract the second entry in a triplet.