Mantis.Oracle.Submit

Description

Submitting transactions to the oracle.

Documentation

createOracle

Arguments

:: Value
The initial datum.

-> Maybe Word64
The metadata key for the initial datum, if any.

-> Maybe Value
The metadata message, if any.

-> LocalNodeConnectInfo CardanoMode
The connection info for the local node.

-> NetworkId
The network identifier.

-> Oracle
The oracle.

-> AddressAny
The payment address.

-> SigningKey PaymentKey
The signing key for the payment address.

-> Quantity
The Lovelace to be sent to the script.

-> MantisM IO TxId
Action to submit the creation transaction.

Submit the transaction to create an oracle.

deleteOracle

Arguments

:: Value
The datum currently held in the script.

-> Maybe Value
The metadata message, if any.

-> LocalNodeConnectInfo CardanoMode
The connection info for the local node.

-> NetworkId
The network identifier.

-> Oracle
The oracle.

-> AddressAny
The payment address.

-> SigningKey PaymentKey
The signing key for the payment address.

-> Quantity
The Lovelace to be sent to the script.

-> MantisM IO TxId
Action to submit the deletion transaction.

Submit the transaction to delete an oracle.

writeOracle

Arguments

:: Value
The datum currently held in the script.

-> Value
The new datum.

-> Maybe Word64
The metadata key for the new datum, if any.

-> Maybe Value
The metadata message, if any.

-> LocalNodeConnectInfo CardanoMode
The connection info for the local node.

-> NetworkId
The network identifier.

-> Oracle
The oracle.

-> AddressAny
The payment address.

-> SigningKey PaymentKey
The signing key for the payment address.

-> Quantity
The Lovelace to be sent to the script.

-> MantisM IO TxId
Action to submit the writing transaction.

Submit the transaction to write new data to an oracle.

operateOracle

Arguments

:: Maybe Action
The redeemer, if any.

-> Maybe Value
The datum currently held in the script, if any.

-> Maybe Value
The new datum, if any.

-> Maybe Word64
The metadata key for the new datum, if any

-> Maybe Value
The metadata message, if any.

-> LocalNodeConnectInfo CardanoMode
The connection info for the local node.

-> NetworkId
The network identifier.

-> Oracle
The oracle.

-> AddressAny
The payment address.

-> SigningKey PaymentKey
The signing key for the payment address.

-> Quantity
The Lovelace to be sent to the script.

-> MantisM IO TxId
Action to submit the deletion transaction.

Submit a transaction to operate the oracle.

build

Arguments

:: Maybe Action
The redeemer, if any.

-> LocalNodeConnectInfo CardanoMode
The connection info for the local node.

-> PlutusScript PlutusScriptV1
The Plutus script.

-> AddressAny
The address of the Plutus script.

-> AddressAny
The payment address.

-> (TxIn, Value)
The UTxO with containing the datum token.

-> (TxIn, Value)
The UTxO containing the control token.

-> [(TxIn, a)]
The UTxOs to be consumed by the transaction.

-> TxIn
The UTxO for collateral.

-> (Maybe Value, Maybe Value)
The existing datum and the datum to replace it, if any.

-> Maybe TxMetadata
The message metadata, if any.

-> MantisM IO (TxBody AlonzoEra)
The action to build the transaction.

Build an oracle transaction.

submit

Arguments

:: LocalNodeConnectInfo CardanoMode
The connection info for the local node.

-> TxBody AlonzoEra
The transaction body.

-> SigningKey PaymentKey
The signing key.

-> MantisM IO TxId
The action to submit the transaction.

Sign and submit a transaction.

findUTxO

Arguments

:: LocalNodeConnectInfo CardanoMode
The connection info for the local node.

-> AddressAny
The address to query.

-> (Value -> Bool)
The condition on values in the UTxO.

-> MantisM IO [(TxIn, TxOut AlonzoEra)]
The action to find the UTxOs.

Find the UTxOs meeting a criterion.

classToId

Arguments

:: AssetClass
The asset class.

-> MantisM IO AssetId
The asset identifier.

Convert a Plutus AssetClass to a Cardano AssetId.

foistMantisEitherIO'

Arguments

:: Show e

=> Show e'

=> IO (Either e (Either e' a))
The action with the nested Eithers.

-> MantisM IO a
The hoisted action.

Hoist doublely nested Eithers.