Mantis.Oracle.Submit

Description

Submitting transactions to the oracle.

Operations

createOracle

Arguments

:: Value
The initial datum.

-> Maybe Word64
The metadata key for the initial datum, if any.

-> Maybe Value
The metadata message, if any.

-> LocalNodeConnectInfo CardanoMode
The connection info for the local node.

-> NetworkId
The network identifier.

-> Oracle
The oracle.

-> AddressAny
The payment address.

-> SigningKey PaymentKey
The signing key for the payment address.

-> Quantity
The Lovelace to be sent to the script.

-> MantisM IO TxId
Action to submit the creation transaction.

Submit the transaction to create an oracle. The payment address must have *separate* eUTxOs containing the datum token and the control token; it also must have at least one other UTxO containing no tokens.

deleteOracle

Arguments

:: Value
The datum currently held in the script.

-> Maybe Value
The metadata message, if any.

-> LocalNodeConnectInfo CardanoMode
The connection info for the local node.

-> NetworkId
The network identifier.

-> Oracle
The oracle.

-> AddressAny
The payment address.

-> SigningKey PaymentKey
The signing key for the payment address.

-> Quantity
The Lovelace to be sent to the script.

-> MantisM IO TxId
Action to submit the deletion transaction.

Submit the transaction to delete an oracle. The payment address must have an eUTxO containing the control token; it also must have at least one other UTxO containing no tokens.

writeOracle

Arguments

:: Value
The datum currently held in the script.

-> Value
The new datum.

-> Maybe Word64
The metadata key for the new datum, if any.

-> Maybe Value
The metadata message, if any.

-> LocalNodeConnectInfo CardanoMode
The connection info for the local node.

-> NetworkId
The network identifier.

-> Oracle
The oracle.

-> AddressAny
The payment address.

-> SigningKey PaymentKey
The signing key for the payment address.

-> Quantity
The Lovelace to be sent to the script.

-> MantisM IO TxId
Action to submit the writing transaction.

Submit the transaction to write new data to an oracle. The payment address must have an eUTxO containing the control token; it also must have at least one other UTxO containing no tokens.